fixing bnet/realid
if the sender is a full name like "Jennifer Clarke" instead of
"Jenn#1884" then do a reverse
  lookup to be consistently only using
Btag names via the "name" attribute.
  This is to protect the identity
of my customers to ensure that nobody sees
  their full name except for
people who already had access to this information;
  instead of seeing
the full name of the player, they will just see the battletag nickname
"Jenn#1884".